test/test_import_data.py
import pytest
from import_data import (
    process_table_with_limit,
    generate_unique_id,
    check_for_duplicates,
    filter_selected_kits,
    process_ancestry,
    process_ftdna,
    process_mh,
    insert_person,
)

# Mock necessary classes and functions
from unittest.mock import MagicMock

PersonTable = MagicMock()
session = MagicMock()


def test_generate_unique_id():
    assert len(generate_unique_id()) == 36
    assert generate_unique_id("test") != generate_unique_id("test2")


def test_process_table_with_limit():
    # Basic test to check if function returns a list
    mock_data = [1, 2, 3]
    result = process_table_with_limit(session, None, [1, 2, 3], lambda x: x, 2)
    assert isinstance(result, list)
    assert len(result) == 2


def test_check_for_duplicates(mocker):
    mocker.patch('import_data.PersonTable', autospec=True)

    # Test creating a new person
    result = check_for_duplicates(session, "unique_id")
    assert isinstance(result, PersonTable)

    # Test updating an existing person
    result = check_for_duplicates(session, "unique_id")  # Assuming person already exists
    assert isinstance(result, PersonTable)


def test_filter_selected_kits(mocker):
    mocker.patch('import_data.filter_session.query')
    # Basic test to check if function returns a dictionary
    result = filter_selected_kits(session, [])
    assert isinstance(result, dict)


def test_process_ancestry(mocker):
    mocker.patch('import_data.process_table_with_limit')
    # Basic test to check if function returns a list
    result = process_ancestry(session, {})
    assert isinstance(result, list)


def test_process_ftdna(mocker):
    mocker.patch('import_data.process_table_with_limit')
    # Basic test to check if function returns a list
    result = process_ftdna(session, {})
    assert isinstance(result, list)


def test_process_mh(mocker):
    mocker.patch('import_data.process_table_with_limit')
    # Basic test to check if function returns a list
    result = process_mh(session, {})
    assert isinstance(result, list)


def test_insert_person(mocker):
    mocker.patch('import_data.check_for_duplicates')
    # Basic test to check if function doesn't raise exceptions
    insert_person(session, [])


# Add more tests for other functions as needed

if __name__ == '__main__':
    pytest.main()
